Replace completion blocks with a home grown Promise implementation

+++ Allihoopa/Drop/SecondarySharingServiceButton.m
@@ -14,10 +14,10 @@ static NSString* const kTwitter = @"TWITTER";
 	UIColor* color;
 
 	if ([_accountType isEqualToString:kFacebook]) {
-		color = [UIColor colorWithRed:0.24 green:0.35 blue:0.59 alpha:1.0];
+		color = [UIColor colorWithRed:(CGFloat)0.24 green:(CGFloat)0.35 blue:(CGFloat)0.59 alpha:1.0];
 	}
 	else if ([_accountType isEqualToString:kTwitter]) {
-		color = [UIColor colorWithRed:0.00 green:0.67 blue:0.93 alpha:1.0];
+		color = [UIColor colorWithRed:(CGFloat)0.00 green:(CGFloat)0.67 blue:(CGFloat)0.93 alpha:1.0];
 	}
 	else {
 		NSAssert(NO, @"Unsupported account type");
@@ -32,8 +32,8 @@ static NSString* const kTwitter = @"TWITTER";
 	}
 	else {
 		self.backgroundColor = [UIColor clearColor];
-		self.layer.borderColor = [UIColor colorWithWhite:0.61 alpha:1.0].CGColor;
-		self.imageView.tintColor = [UIColor colorWithWhite:0.61 alpha:1.0];
+		self.layer.borderColor = [UIColor colorWithWhite:(CGFloat)0.61 alpha:1.0].CGColor;
+		self.imageView.tintColor = [UIColor colorWithWhite:(CGFloat)0.61 alpha:1.0];
 	}
 
 	self.layer.cornerRadius = self.frame.size.width / 2;
